Finds markers that are conserved between the groups

Finds markers that are conserved between the groups

FindConservedMarkers(
  object,
  ident.1,
  ident.2 = NULL,
  grouping.var,
  assay = "RNA",
  slot = "data",
  meta.method = minimump,
  verbose = TRUE,
  ...
)

Arguments

object
An object

ident.1
Identity class to define markers for

ident.2
A second identity class for comparison. If NULL (default) - use all other cells for comparison.

grouping.var
grouping variable

assay
of assay to fetch data for (default is RNA)

slot
Slot to pull data from; note that if test.use is "negbinom", "poisson", or "DESeq2", slot will be set to "counts"

meta.method
method for combining p-values. Should be a function from the metap package (NOTE: pass the function, not a string)

verbose
Print a progress bar once expression testing begins

...
parameters to pass to FindMarkers

Value

data.frame containing a ranked list of putative conserved markers, and associated statistics (p-values within each group and a combined p-value (such as Fishers combined p-value or others from the metap package), percentage of cells expressing the marker, average differences). Name of group is appended to each associated output column (e.g. CTRL_p_val). If only one group is tested in the grouping.var, max and combined p-values are not returned.

Examples

if (FALSE) {
pbmc_small
# Create a simulated grouping variable
pbmc_small[['groups']] <- sample(x = c('g1', 'g2'), size = ncol(x = pbmc_small), replace = TRUE)
FindConservedMarkers(pbmc_small, ident.1 = 0, ident.2 = 1, grouping.var = "groups")
}