Query Antibodies from multiple vendors

Queries the respective antibody for each vendor and when posible filters for antibodies for IFC, IHC, and FLOW.

query_cc_antibodies(search_term, sleep = 1)

query_sc_antibodies(search_term, sleep = 1)

query_biocompare_antibodies(search_term, sleep = 1)

query_biolegend_antibodies(search_term, sleep = 1)

Arguments

search_term
The term used to query the antibodies

sleep
an ammount of time to wait before submitting the query. Check details for more information

Value

NULL when no results are found or a DF with information

Details

Most of these functions have been implemented as queries to the vendor website, therefore if many queries are done in quick succession, it is posible that the vendor considers the queries as an attack to their website. Therefore we have the sleep argument set to 1 by default, the function will wait 1 second before doing the query. If many queries will be done, feel free to modify the parameter but be aware that some requests might be denied.

Examples

query_cc_antibodies("CD11bfakename")

#> NULL

# NULL
query_cc_antibodies("CD11c")

#>    Cat_num                                       Product Name Reactivity
#> 1    45581                       CD11c (D3V1E) XP® Rabbit mAb          H
#> 2    97585                           CD11c (D1V9Y) Rabbit mAb          M
#> 3    97473 CD11c (3.9) Mouse mAb (violetFluor™ 450 Conjugate)          H
#> 4    42847    CD11c (3.9) Mouse mAb (redFluor™ 710 Conjugate)          H
#> 5    93983  CD11c (N418) Hamster mAb (PerCP-Cy5.5® Conjugate)          M
#> 6    77971       CD11c (N418) Hamster mAb (PE-Cy7® Conjugate)          M
#> 7    80342          CD11c (3.9) Mouse mAb (PE-Cy7® Conjugate)          H
#> 8    56025               CD11c (3.9) Mouse mAb (PE Conjugate)          H
#> 9    73359            CD11c (N418) Hamster mAb (PE Conjugate)          M
#> 10   51419          CD11c (N418) Hamster mAb (FITC Conjugate)          M
#> 11   69627             CD11c (3.9) Mouse mAb (FITC Conjugate)          H
#> 12   37664      CD11c (N418) Hamster mAb (APC-Cy7® Conjugate)          M
#> 13   51562         CD11c (3.9) Mouse mAb (APC-Cy7® Conjugate)          H
#> 14   33293           CD11c (N418) Hamster mAb (APC Conjugate)          M
#> 15   36268              CD11c (3.9) Mouse mAb (APC Conjugate)          H
#> 16   62824                              CD11c (3.9) Mouse mAb          H
#>                   Product_Type
#> 1  COMPARE  Primary Antibodies
#> 2  COMPARE  Primary Antibodies
#> 3  COMPARE Antibody Conjugates
#> 4  COMPARE Antibody Conjugates
#> 5  COMPARE Antibody Conjugates
#> 6  COMPARE Antibody Conjugates
#> 7  COMPARE Antibody Conjugates
#> 8  COMPARE Antibody Conjugates
#> 9  COMPARE Antibody Conjugates
#> 10 COMPARE Antibody Conjugates
#> 11 COMPARE Antibody Conjugates
#> 12 COMPARE Antibody Conjugates
#> 13 COMPARE Antibody Conjugates
#> 14 COMPARE Antibody Conjugates
#> 15 COMPARE Antibody Conjugates
#> 16 COMPARE  Primary Antibodies

#   Cat_num                                       Product Name Reactivity
# 1   97585                           CD11c (D1V9Y) Rabbit mAb          M
# 2   97473 CD11c (3.9) Mouse mAb (violetFluor™ 450 Conjugate)          H
# 3   42847    CD11c (3.9) Mouse mAb (redFluor™ 710 Conjugate)          H
# 4   80342          CD11c (3.9) Mouse mAb (PE-Cy7® Conjugate)          H
# 5   56025               CD11c (3.9) Mouse mAb (PE Conjugate)          H
# 6   69627             CD11c (3.9) Mouse mAb (FITC Conjugate)          H
# 7   51562         CD11c (3.9) Mouse mAb (APC-Cy7® Conjugate)          H
# 8   36268              CD11c (3.9) Mouse mAb (APC Conjugate)          H
# Product_Type
# 1  Primary Antibodies
# 2 Antibody Conjugates
# 3 Antibody Conjugates
# 4 Antibody Conjugates
# 5 Antibody Conjugates
# 6 Antibody Conjugates
# 7 Antibody Conjugates
# 8 Antibody Conjugates

query_sc_antibodies("CD11bfakename")

#> NULL

# NULL
head(query_sc_antibodies("CD11C"))

#>                        Product_Name  Cat_num Citations Rating     Epitope
#> 1   Integrin alpha X Antibody (B-6) sc-46676         3    (8) Integrin αX
#> 2   Integrin alpha X Antibody (3.9)  sc-1185         2    (2) Integrin αX
#> 3 Integrin alpha X Antibody (B-Iy6) sc-19989         1    (1) Integrin αX
#> 4  Integrin alpha X Antibody (N418) sc-23951         3    (2) Integrin αX
#> 6 Integrin alpha X Antibody (2Q862) sc-71454       NEW    (1) Integrin αX
#> 8 Integrin alpha X Antibody (2Q865) sc-71455       NEW    (1) Integrin αX
#>   Species                       Method
#> 1   human WB, IP, IF, IHC(P) and ELISA
#> 2   human               IP, IF and FCM
#> 3   human                   IF and FCM
#> 4   mouse                   IF and FCM
#> 6   human                   IF and FCM
#> 8   human                   IP and FCM

# Product_Name  Cat_num Citations Rating     Epitope
# 1   Integrin alpha X Antibody (B-6) sc-46676         3    (8) Integrin aX
# 2   Integrin alpha X Antibody (3.9)  sc-1185         2    (2) Integrin aX
# 3 Integrin alpha X Antibody (B-Iy6) sc-19989         1    (1) Integrin aX
# 4  Integrin alpha X Antibody (N418) sc-23951         3    (2) Integrin aX
# 6 Integrin alpha X Antibody (2Q865) sc-71455       NEW    (1) Integrin aX
# 7 Integrin alpha X Antibody (3H986) sc-71456       NEW    (1) Integrin aX
# Species                       Method
# 1   human WB, IP, IF, IHC(P) and ELISA
# 2   human               IP, IF and FCM
# 3   human                   IF and FCM
# 4   mouse                   IF and FCM
# 6   human                   IP and FCM
# 7   mouse                   IF and FCM

#' query_biocompare_antibodies("CD11bfakename")
# NULL
head(query_biocompare_antibodies("CD11C"),3)

#>                                                                          title
#> 1 Anti-Human CD11c Monoclonal Antibody Unconjugated Conjugated, Flow Validated
#> 2                       Anti-CD11c (Dendritic Cell Marker) Monoclonal Antibody
#> 3                                        Anti-CD11c Rabbit Monoclonal Antibody
#>      vendor
#> 1 BosterBio
#> 2 BosterBio
#> 3 BosterBio
#>                                                                                                      specification
#> 1 Applications: Flow Cytometry (FCM); Reactivity: Human (Hu); Conjugate/Tag: Unconjugated; Quantity: Various Sizes
#> 2    Applications: Flow Cytometry (FCM); Reactivity: Human (Hu); Conjugate/Tag: Unconjugated; Quantity: 100ug/vial
#> 3    Applications: WB, ICC, IF, IHC, IP; Reactivity: Human (Hu); Conjugate/Tag: Unconjugated; Quantity: 100ug/vial

query_biolegend_antibodies("CD11bfakename")

#> character(0)

# NULL
head(query_biolegend_antibodies("CD11C"))

#> [1] "MojoSort™ Mouse CD11c Nanobeads"  "APC anti-human CD11c Antibody"   
#> [3] "Biotin anti-human CD11c Antibody" "FITC anti-human CD11c Antibody"  
#> [5] "PE anti-human CD11c Antibody"     "PE/Cy5 anti-human CD11c Antibody"

# [1] "MojoSortâ\u0084¢ Mouse CD11c Nanobeads"
# [2] "APC anti-human CD11c Antibody"
# [3] "Biotin anti-human CD11c Antibody"
# [4] "FITC anti-human CD11c Antibody"
# [5] "PE anti-human CD11c Antibody"
# [6] "PE/Cy5 anti-human CD11c Antibody"