Fits a decision tree in one data set and tests the performance in another

Fits a decision tree in one data set and tests the performance in another

cross_validate(
  train,
  test,
  cluster,
  genes_use = Seurat::VariableFeatures(train),
  warn.gene.removal = TRUE,
  ...
)

Arguments

train
a Seurat object to be used for trainning.

test
another Seurat object to be used for testing.

cluster
the cluster whose equivalence needs to be found.

genes_use
character vector specifying which genes to use for the classification, defaults to Seurat::VariableFeatures(train)

warn.gene.removal
logical indicating wether to warn the user when genes are removed because they are missing in one of the datasets. defults to TRUE

...
additional arguments to be passed to ranger_importances.Seurat

Value

a list containing the (1) tree fit, (2) a summary_table (3) the concensus rules of the tree (4) ranger_significance_table (5) the suggested genes for the gating

Examples

cross_validate(small_5050_mix, small_9901_mix, cluster = "0")

#> Warning: Only few negative importance values found, inaccurate p-values. Consider the 'altmann' approach.
#> This can be done by setting the argument 'imp_method' to 'altmann', note that this method is extremely computationally intensive.
#> For more information please refer to ?ranger::ranger

#> Warning: Some important genes were removed because they are not present in the test dataset. 
#> Removed genes: ASNS, CD3D, HEY1, ADA, RPL26, XIST, CA2, TSC22D3, CDKN2A, AIF1, GAL, MDK, MAP1A, PSMB8, PSMB9, FAM127B, DMKN, TSTD1, PYGL, IFI16, KRT18, CSRP2, SPRED2, ZNF503, ZFAT, PRKAG2-AS1, CDC42EP1, CTC1, LBX1-AS1, HSPA1B, GTF2IRD1, HOXA9

#> $party_fit
#> 
#> Model formula:
#> ident ~ ARHGDIB + TMSB4X + MZB1 + SOX4 + FYB + UBE2C + CD1E + 
#>     HIST1H1E + ITGA4 + HIST1H4C + CXCR4 + CDK1 + CHI3L2 + ITM2A + 
#>     ID3 + MALAT1 + JUN + C12orf57 + HIST1H1C + MYC + DDIT4 + 
#>     HIST1H2BK + SH3TC1
#> 
#> Fitted party:
#> [1] root
#> |   [2] ARHGDIB <= 2.29381
#> |   |   [3] SOX4 <= 2.74241: not clus 0 (n = 90, err = 27.8%)
#> |   |   [4] SOX4 > 2.74241: clus 0 (n = 8, err = 0.0%)
#> |   [5] ARHGDIB > 2.29381
#> |   |   [6] DDIT4 <= 2.13797: clus 0 (n = 148, err = 4.1%)
#> |   |   [7] DDIT4 > 2.13797: clus 0 (n = 9, err = 44.4%)
#> 
#> Number of inner nodes:    3
#> Number of terminal nodes: 4
#> 
#> $confusion_matrix
#>             cluster
#> predicted      0   1   2
#>   clus 0     263  98   8
#>   not clus 0   0   0  15
#> 
#> $concensus_rules
#> Cluster-clus 0: 
#> 	all elements:
#> 		ARHGDIB +
#> 	majority elements:
#> 		DDIT4 -
#> Cluster-not clus 0: 
#> 	all elements:
#> 		ARHGDIB -
#> 		SOX4 -
#> 
#> $ranger_significance_table
#>            importance     pvalue       gene
#> ASNS        6.0094119 0.00000000       ASNS
#> ARHGDIB     5.1676992 0.00000000    ARHGDIB
#> CD3D        4.8032257 0.00000000       CD3D
#> TMSB4X      4.3671933 0.00000000     TMSB4X
#> HEY1        3.4772237 0.00000000       HEY1
#> ADA         3.1096150 0.00000000        ADA
#> MZB1        2.3375299 0.00000000       MZB1
#> RPL26       2.3216236 0.00000000      RPL26
#> XIST        2.0958065 0.00000000       XIST
#> CA2         1.9730605 0.00000000        CA2
#> SOX4        1.9680173 0.00000000       SOX4
#> TSC22D3     1.8654082 0.00000000    TSC22D3
#> CDKN2A      1.7728410 0.00000000     CDKN2A
#> AIF1        1.6844896 0.00000000       AIF1
#> FYB         1.6233718 0.00000000        FYB
#> GAL         1.1649316 0.00000000        GAL
#> MDK         1.0080020 0.00000000        MDK
#> UBE2C       0.9583872 0.00000000      UBE2C
#> MAP1A       0.8532897 0.00000000      MAP1A
#> CD1E        0.7916297 0.00000000       CD1E
#> HIST1H1E    0.7725623 0.00000000   HIST1H1E
#> PSMB8       0.7516684 0.00000000      PSMB8
#> PSMB9       0.6452038 0.00000000      PSMB9
#> FAM127B     0.6252417 0.00000000    FAM127B
#> DMKN        0.6163686 0.00000000       DMKN
#> TSTD1       0.5738816 0.00000000      TSTD1
#> ITGA4       0.5449013 0.00000000      ITGA4
#> HIST1H4C    0.5237232 0.00000000   HIST1H4C
#> CXCR4       0.4811929 0.00000000      CXCR4
#> CDK1        0.4179864 0.00000000       CDK1
#> CHI3L2      0.3775718 0.00000000     CHI3L2
#> PYGL        0.3621761 0.00000000       PYGL
#> IFI16       0.2980396 0.00000000      IFI16
#> ITM2A       0.2636847 0.00000000      ITM2A
#> KRT18       0.2550561 0.00000000      KRT18
#> CSRP2       0.2460805 0.00000000      CSRP2
#> ID3         0.2459439 0.00000000        ID3
#> MALAT1      0.2277814 0.01219512     MALAT1
#> JUN         0.2170782 0.02439024        JUN
#> SPRED2      0.2093111 0.02439024     SPRED2
#> ZNF503      0.1942172 0.02439024     ZNF503
#> ZFAT        0.1916341 0.02439024       ZFAT
#> PRKAG2-AS1  0.1863027 0.02439024 PRKAG2-AS1
#> CDC42EP1    0.1731541 0.02439024   CDC42EP1
#> CTC1        0.1710720 0.02439024       CTC1
#> C12orf57    0.1683163 0.02439024   C12orf57
#> LBX1-AS1    0.1675656 0.02439024   LBX1-AS1
#> HIST1H1C    0.1628978 0.02439024   HIST1H1C
#> HSPA1B      0.1553756 0.02439024     HSPA1B
#> MYC         0.1537134 0.03658537        MYC
#> GTF2IRD1    0.1483967 0.04878049   GTF2IRD1
#> HOXA9       0.1471439 0.04878049      HOXA9
#> DDIT4       0.1449510 0.04878049      DDIT4
#> HIST1H2BK   0.1363704 0.04878049  HIST1H2BK
#> SH3TC1      0.1304120 0.04878049     SH3TC1
#> 
#> $gating_genes
#> [1] "ARHGDIB" "SOX4"    "DDIT4"  
#> 

cross_validate(small_5050_mix, small_9901_mix, cluster = "ALL")

#> Warning: Only few negative importance values found, inaccurate p-values. Consider the 'altmann' approach.
#> This can be done by setting the argument 'imp_method' to 'altmann', note that this method is extremely computationally intensive.
#> For more information please refer to ?ranger::ranger

#> Warning: Some important genes were removed because they are not present in the test dataset. 
#> Removed genes: ASNS, CD3D, ADA, HEY1, TSC22D3, RPL26, XIST, AIF1, MAP1A, CDKN2A, CA2, PSMB8, FAM127B, TSTD1, GAL, MDK, DMKN, CSRP2, ID2, KRT18, PSMB9, HSPA1B

#> $party_fit
#> 
#> Model formula:
#> ident ~ TMSB4X + ARHGDIB + MZB1 + SOX4 + FYB + CD1E + UBE2C + 
#>     HIST1H1E + HIST1H4C + ITM2A + IGLL1 + CXCR4 + CHI3L2 + HIST1H2BK + 
#>     ITGA4 + DDIT4 + C12orf57 + HIST1H1C
#> 
#> Fitted party:
#> [1] root
#> |   [2] ARHGDIB <= 2.29381
#> |   |   [3] SOX4 <= 2.74241: 1 (n = 90, err = 27.8%)
#> |   |   [4] SOX4 > 2.74241: 0 (n = 8, err = 0.0%)
#> |   [5] ARHGDIB > 2.29381
#> |   |   [6] DDIT4 <= 2.13797: 0 (n = 148, err = 4.1%)
#> |   |   [7] DDIT4 > 2.13797: 0 (n = 9, err = 44.4%)
#> 
#> Number of inner nodes:    3
#> Number of terminal nodes: 4
#> 
#> $confusion_matrix
#>          cluster
#> predicted   0   1   2
#>         0 263  98   8
#>         1   0   0  15
#> 
#> $concensus_rules
#> Cluster-0: 
#> 	all elements:
#> 		ARHGDIB +
#> 	majority elements:
#> 		DDIT4 -
#> Cluster-1: 
#> 	all elements:
#> 		ARHGDIB -
#> 		SOX4 -
#> 
#> $ranger_significance_table
#>           importance     pvalue      gene
#> ASNS       5.6354842 0.00000000      ASNS
#> CD3D       4.8789549 0.00000000      CD3D
#> TMSB4X     4.3761890 0.00000000    TMSB4X
#> ARHGDIB    4.2833060 0.00000000   ARHGDIB
#> MZB1       3.4198224 0.00000000      MZB1
#> ADA        3.1530134 0.00000000       ADA
#> HEY1       2.8489714 0.00000000      HEY1
#> TSC22D3    2.1903904 0.00000000   TSC22D3
#> RPL26      2.1838063 0.00000000     RPL26
#> XIST       2.1728741 0.00000000      XIST
#> AIF1       2.0274626 0.00000000      AIF1
#> SOX4       1.7082333 0.00000000      SOX4
#> MAP1A      1.6491278 0.00000000     MAP1A
#> FYB        1.4254269 0.00000000       FYB
#> CDKN2A     1.3816793 0.00000000    CDKN2A
#> CA2        1.3786192 0.00000000       CA2
#> CD1E       0.9548178 0.00000000      CD1E
#> UBE2C      0.9294106 0.00000000     UBE2C
#> PSMB8      0.8618613 0.00000000     PSMB8
#> FAM127B    0.7963698 0.00000000   FAM127B
#> HIST1H1E   0.7267816 0.00000000  HIST1H1E
#> TSTD1      0.7098676 0.00000000     TSTD1
#> GAL        0.6304936 0.00000000       GAL
#> HIST1H4C   0.6114589 0.00000000  HIST1H4C
#> MDK        0.5016790 0.00000000       MDK
#> ITM2A      0.4967601 0.00000000     ITM2A
#> DMKN       0.4876830 0.00000000      DMKN
#> CSRP2      0.4863419 0.00000000     CSRP2
#> IGLL1      0.4115364 0.00000000     IGLL1
#> ID2        0.3743618 0.01041667       ID2
#> CXCR4      0.3713671 0.01041667     CXCR4
#> KRT18      0.3702754 0.01041667     KRT18
#> PSMB9      0.3702514 0.01041667     PSMB9
#> CHI3L2     0.3552276 0.01041667    CHI3L2
#> HIST1H2BK  0.3164331 0.02083333 HIST1H2BK
#> ITGA4      0.3039233 0.02083333     ITGA4
#> DDIT4      0.2642464 0.04166667     DDIT4
#> HSPA1B     0.2380191 0.04166667    HSPA1B
#> C12orf57   0.2359516 0.04166667  C12orf57
#> HIST1H1C   0.2312922 0.04166667  HIST1H1C
#> 
#> $gating_genes
#> [1] "ARHGDIB" "SOX4"    "DDIT4"  
#>