Fits a classification tree on a Seurat object

Fits a classification tree on a Seurat object

fit_ctree(
  object,
  genes_use = Seurat::VariableFeatures(object),
  cluster = "ALL",
  ...
)

Arguments

object
a Seurat object

genes_use
a character vector indicating which genes to use in the classification. currently implemented only for Seurat objects. (for data frames one can simply subset the input data frame) defaults to Seurat::VariableFeatures(object)

cluster
a cluster name for which the markers will be found

...
additional arguments to be passed to partykit::ctree_control

Value

a ctree fit

Examples

   fit_ctree(small_9901_mix, c('CCNB1', 'PLK1', 'AURKA'), cluster = 'ALL')

#> 
#> Model formula:
#> ident ~ CCNB1 + PLK1 + AURKA
#> 
#> Fitted party:
#> [1] root
#> |   [2] PLK1 <= 2.31327
#> |   |   [3] CCNB1 <= 3.01197
#> |   |   |   [4] AURKA <= 2.90958
#> |   |   |   |   [5] PLK1 <= 1.75363: 0 (n = 199, err = 3.5%)
#> |   |   |   |   [6] PLK1 > 1.75363: 0 (n = 41, err = 17.1%)
#> |   |   |   [7] AURKA > 2.90958: 0 (n = 7, err = 42.9%)
#> |   |   [8] CCNB1 > 3.01197
#> |   |   |   [9] PLK1 <= 1.67634: 0 (n = 26, err = 38.5%)
#> |   |   |   [10] PLK1 > 1.67634: 1 (n = 23, err = 26.1%)
#> |   [11] PLK1 > 2.31327
#> |   |   [12] AURKA <= 2.02766: 1 (n = 21, err = 47.6%)
#> |   |   [13] AURKA > 2.02766: 1 (n = 67, err = 19.4%)
#> 
#> Number of inner nodes:    6
#> Number of terminal nodes: 7

   fit_ctree(small_9901_mix, c('CCNB1', 'PLK1', 'AURKA'), cluster = '0')

#> 
#> Model formula:
#> ident ~ CCNB1 + PLK1 + AURKA
#> 
#> Fitted party:
#> [1] root
#> |   [2] PLK1 <= 2.31327
#> |   |   [3] CCNB1 <= 3.01197
#> |   |   |   [4] PLK1 <= 1.75363: clus 0 (n = 202, err = 4.0%)
#> |   |   |   [5] PLK1 > 1.75363: clus 0 (n = 45, err = 20.0%)
#> |   |   [6] CCNB1 > 3.01197
#> |   |   |   [7] PLK1 <= 1.67634: clus 0 (n = 26, err = 38.5%)
#> |   |   |   [8] PLK1 > 1.67634: not clus 0 (n = 23, err = 17.4%)
#> |   [9] PLK1 > 2.31327
#> |   |   [10] AURKA <= 2.02766: not clus 0 (n = 21, err = 47.6%)
#> |   |   [11] AURKA > 2.02766: not clus 0 (n = 67, err = 4.5%)
#> 
#> Number of inner nodes:    5
#> Number of terminal nodes: 6

   
# 
# Model formula:
# ident ~ CCNB1 + PLK1 + AURKA
# 
# Fitted party:
# [1] root
# |   [2] PLK1 <= 2.31327
# |   |   [3] CCNB1 <= 3.01197
# |   |   |   [4] PLK1 <= 1.75363: clus 0 (n = 202, err = 4.0%)
# |   |   |   [5] PLK1 > 1.75363: clus 0 (n = 45, err = 20.0%)
# |   |   [6] CCNB1 > 3.01197
# |   |   |   [7] PLK1 <= 1.67634: clus 0 (n = 26, err = 38.5%)
# |   |   |   [8] PLK1 > 1.67634: not clus 0 (n = 23, err = 17.4%)
# |   [9] PLK1 > 2.31327
# |   |   [10] AURKA <= 2.02766: not clus 0 (n = 21, err = 47.6%)
# |   |   [11] AURKA > 2.02766: not clus 0 (n = 67, err = 4.5%)
# 
# Number of inner nodes:    5
# Number of terminal nodes: 6