ranger_importances

Calculates ranger-based variable importances for data frames and Seurat objects

ranger_importances(x, ...)

FindAllMarkers_ranger.Seurat(
  object,
  genes_use = Seurat::VariableFeatures(object),
  ...
)

Arguments

...
additional arguments to be passed to ranger

object
a Seurat or data frame object

genes_use
a character vector indicating which genes to use in the classification. currently implemented only for Seurat objects. (for data frames one can simply subset the input data frame)

cluster
a cluster name for which the markers will be found

pval_cutoff
p value cutoff for the markers

imp_method
importance method, either of "janitza" or "altmann"

num.trees
number of trees to be build using ranger

return_what
a subset of "ranger_fit", "importances_ranger", "signif_importances_ranger", defaults to signif_importances_ranger

warn.imp.method
logical indicating wether warning should be issued when few negative importances are found to calculate the p.values in ranger.

return
ranger_fit, importances_ranger, signif_importances_ranger

Value

by default returns a data frame with the importances and p values but this behavior can be modified by the

Functions

FindAllMarkers_ranger.Seurat: Calculate variable importances to each cluster in a Seurat object

Examples

    head(ranger_importances(Seurat::pbmc_small, cluster = 'ALL', warn.imp.method = FALSE))

#>          importance pvalue     gene
#> HLA-DPB1  8.2824862      0 HLA-DPB1
#> S100A8    7.4094517      0   S100A8
#> S100A9    6.1534456      0   S100A9
#> HLA-DQA1  3.0576261      0 HLA-DQA1
#> GNLY      1.6808031      0     GNLY
#> VDAC3     0.4500117      0    VDAC3

    
#          importance pvalue     gene
# HLA-DPB1   8.226410      0 HLA-DPB1
# S100A9     6.648748      0   S100A9
# S100A8     6.537872      0   S100A8
# HLA-DQA1   3.103721      0 HLA-DQA1
# GNLY       1.779237      0     GNLY