ranger_importances

Calculates ranger-based variable importances for data frames and Seurat objects

ranger_importances.df(
  object,
  cluster = NULL,
  pval_cutoff = 0.05,
  imp_method = c("janitza", "altmann"),
  num.trees = 500,
  warn.imp.method = TRUE,
  identity_col_name = "ident",
  return_what = "signif_importances_ranger",
  sort_results = TRUE,
  ...
)

ranger_importances.Seurat(
  object,
  cluster = NULL,
  pval_cutoff = 0.05,
  imp_method = c("janitza", "altmann"),
  num.trees = 500,
  genes_use = Seurat::VariableFeatures(object),
  warn.imp.method = TRUE,
  ...
)

FindAllMarkers_ranger.Seurat(
  object,
  genes_use = Seurat::VariableFeatures(object),
  ...
)

Arguments

object
a Seurat or data frame object

cluster
a cluster name for which the markers will be found

pval_cutoff
p value cutoff for the markers

imp_method
importance method, either of "janitza" or "altmann"

num.trees
number of trees to be build using ranger

warn.imp.method
logical indicating wether warning should be issued when few negative importances are found to calculate the p.values in ranger.

return_what
a subset of "ranger_fit", "importances_ranger", "signif_importances_ranger", defaults to signif_importances_ranger

...
additional arguments to be passed to ranger

genes_use
a character vector indicating which genes to use in the classification. currently implemented only for Seurat objects. (for data frames one can simply subset the input data frame)

return
ranger_fit, importances_ranger, signif_importances_ranger

Value

by default returns a data frame with the importances and p values but this behavior can be modified by the

Functions

ranger_importances.Seurat: Calculate variable importances to calssify a Seurat object

FindAllMarkers_ranger.Seurat: Calculate variable importances to each cluster in a Seurat object

Examples

    head(ranger_importances.Seurat(Seurat::pbmc_small, cluster = 'ALL', warn.imp.method = FALSE))

#>          importance pvalue     gene
#> HLA-DPB1  8.4808212      0 HLA-DPB1
#> S100A8    6.9268778      0   S100A8
#> S100A9    5.9730276      0   S100A9
#> HLA-DQA1  2.8349639      0 HLA-DQA1
#> GNLY      1.5144386      0     GNLY
#> PPBP      0.4182264      0     PPBP

    
#          importance pvalue     gene
# HLA-DPB1  7.7450251      0 HLA-DPB1
# S100A8    6.8548703      0   S100A8
# S100A9    6.3483212      0   S100A9
# HLA-DQA1  2.9561073      0 HLA-DQA1
# GNLY      2.0354203      0     GNLY
# VDAC3     0.6537556      0    VDAC3